pyproject.toml
[project]
name = "taxi-lakehouse-ai-agent"
version = "0.1.0"
description = "Basic repository scaffold for a taxi lakehouse and AI query agent."
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.1.0",
  "fastapi>=0.115.0",
  "openai>=1.75.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.2",
  "sqlglot>=25.20.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
